requirements.txt
streamlit==1.24.0
Pillow==9.5.0
numpy==1.23.5
opencv-python-headless==4.8.0.74
torch==2.0.1
ultralytics==8.0.20
